Automated ML pipeline/requirements.txt
google-cloud-storage==2.5.0
google-cloud-aiplatform==1.13.0
scikit-learn==1.0.2
pandas==1.4.2
numpy==1.22.3
xgboost==1.6.1
tensorflow==2.9.1
pyyaml==6.0
pytest==7.1.2